Pass a TimestampWithTimezone into scheduler methods rather than a pendulum datetime
Summary:
Inching towards pendulum removal by passing around a different object.

Test Plan: BK
