+++ python_modules/dagster/dagster/_api/snapshot_schedule.py
@@ -1,14 +1,14 @@
-from typing import TYPE_CHECKING, Any, Optional, Sequence
+from typing import TYPE_CHECKING, Optional, Sequence
 
 import dagster._check as check
 from dagster._core.definitions.schedule_definition import ScheduleExecutionData
+from dagster._core.definitions.timestamp import TimestampWithTimezone
 from dagster._core.errors import DagsterUserCodeProcessError
 from dagster._core.instance import DagsterInstance
 from dagster._core.remote_representation.external_data import ExternalScheduleExecutionErrorData
 from dagster._core.remote_representation.handle import RepositoryHandle
 from dagster._grpc.types import ExternalScheduleExecutionArgs
 from dagster._serdes import deserialize_value
-from dagster._seven.compat.pendulum import PendulumDateTime
 
 if TYPE_CHECKING:
     from dagster._grpc.client import DagsterGrpcClient
@@ -18,7 +18,7 @@ def sync_get_external_schedule_execution_data_ephemeral_grpc(
     instance: DagsterInstance,
     repository_handle: RepositoryHandle,
     schedule_name: str,
-    scheduled_execution_time: Any,
+    scheduled_execution_time: Optional[TimestampWithTimezone],
     log_key: Optional[Sequence[str]],
     timeout: Optional[int] = None,
 ) -> ScheduleExecutionData:
@@ -44,13 +44,15 @@ def sync_get_external_schedule_execution_data_grpc(
     instance: DagsterInstance,
     repository_handle: RepositoryHandle,
     schedule_name: str,
-    scheduled_execution_time: Any,
+    scheduled_execution_time: Optional[TimestampWithTimezone],
     log_key: Optional[Sequence[str]],
     timeout: Optional[int] = None,
 ) -> ScheduleExecutionData:
     check.inst_param(repository_handle, "repository_handle", RepositoryHandle)
     check.str_param(schedule_name, "schedule_name")
-    check.opt_inst_param(scheduled_execution_time, "scheduled_execution_time", PendulumDateTime)
+    check.opt_inst_param(
+        scheduled_execution_time, "scheduled_execution_time", TimestampWithTimezone
+    )
 
     origin = repository_handle.get_external_origin()
     result = deserialize_value(
@@ -60,10 +62,10 @@ def sync_get_external_schedule_execution_data_grpc(
                 instance_ref=instance.get_ref(),
                 schedule_name=schedule_name,
                 scheduled_execution_timestamp=(
-                    scheduled_execution_time.timestamp() if scheduled_execution_time else None
+                    scheduled_execution_time.timestamp if scheduled_execution_time else None
                 ),
                 scheduled_execution_timezone=(
-                    scheduled_execution_time.timezone.name if scheduled_execution_time else None
+                    scheduled_execution_time.timezone if scheduled_execution_time else None
                 ),
                 log_key=log_key,
                 timeout=timeout,
