+++ python_modules/dagster/dagster/_core/definitions/timestamp.py
@@ -1,0 +1,19 @@
+from typing import NamedTuple
+
+from dagster._serdes import whitelist_for_serdes
+from dagster._seven.compat.pendulum import PendulumDateTime
+
+
+# TimestampWithTimezone is used to preserve IANA timezone information when serializing.
+# Serializing with the UTC offset (i.e. via datetime.isoformat) is insufficient because offsets vary
+# depending on daylight savings time. This causes timedelta operations to be inexact, since the
+# exact timezone is not preserved. To prevent any lossy serialization, ths implementation
+# serializes both the datetime float and the IANA timezone.
+@whitelist_for_serdes
+class TimestampWithTimezone(NamedTuple):
+    timestamp: float  # Seconds since the Unix epoch
+    timezone: str
+
+    @staticmethod
+    def from_pendulum_time(dt: PendulumDateTime):
+        return TimestampWithTimezone(dt.timestamp(), dt.timezone.name)
