+++ python_modules/dagster/dagster/_scheduler/scheduler.py
@@ -27,6 +27,7 @@ import dagster._check as check
 from dagster._core.definitions.run_request import RunRequest
 from dagster._core.definitions.schedule_definition import DefaultScheduleStatus
 from dagster._core.definitions.selector import JobSubsetSelector
+from dagster._core.definitions.timestamp import TimestampWithTimezone
 from dagster._core.definitions.utils import normalize_tags
 from dagster._core.errors import DagsterCodeLocationLoadError, DagsterUserCodeUnreachableError
 from dagster._core.instance import DagsterInstance
@@ -621,6 +622,7 @@ def launch_scheduled_runs_for_schedule_iterator(
                     logger,
                     external_schedule,
                     schedule_time,
+                    timezone_str,
                     tick_context,
                     submit_threadpool_executor,
                     schedule_debug_crash_flags,
@@ -783,6 +785,7 @@ def _schedule_runs_at_time(
     logger: logging.Logger,
     external_schedule: ExternalSchedule,
     schedule_time: datetime.datetime,
+    timezone_str: str,
     tick_context: _ScheduleLaunchContext,
     submit_threadpool_executor: Optional[ThreadPoolExecutor],
     debug_crash_flags: Optional[SingleInstigatorDebugCrashFlags] = None,
@@ -796,7 +799,10 @@ def _schedule_runs_at_time(
         instance=instance,
         repository_handle=repository_handle,
         schedule_name=external_schedule.name,
-        scheduled_execution_time=schedule_time,
+        scheduled_execution_time=TimestampWithTimezone(
+            schedule_time.timestamp(),
+            timezone_str,
+        ),
         log_key=tick_context.log_key,
     )
     yield None
