+++ python_modules/dagster/dagster_tests/core_tests/test_mask_user_code_errors.py
@@ -1,12 +1,13 @@
 import re
 import sys
+import time
 from typing import Any
 
 import pytest
 from dagster import Config, RunConfig, config_mapping, job, op
+from dagster._core.definitions.timestamp import TimestampWithTimezone
 from dagster._core.errors import DagsterUserCodeProcessError
 from dagster._core.test_utils import environ, instance_for_test
-from dagster._seven import get_current_datetime_in_utc
 from dagster._utils.error import serializable_error_info_from_exc_info
 
 from ..api_tests.utils import get_bar_repo_handle
@@ -85,7 +86,7 @@ def test_masking_schedule_execution(instance, enable_masking_user_code_errors, c
                 instance,
                 repository_handle,
                 "schedule_error",
-                get_current_datetime_in_utc(),
+                TimestampWithTimezone(time.time(), "UTC"),
                 None,
                 None,
             )
